+++ src/components/Header.js
@@ -1,0 +1,24 @@
+import React from 'react'
+import { CDN_LOGO } from '../utils/constants'
+const Header = () => {
+    return (
+        <div className="header">
+          <div className="logo-container">
+            <img
+              className="logo"
+              alt="logo"
+              src={CDN_LOGO}
+            />
+          </div>
+          <div className="nav-items">
+            <ul>
+              <li>Home</li>
+              <li>About</li>
+              <li>Contact us</li>
+            </ul>
+          </div>
+        </div>
+      );
+}
+
+export default Header
