+++ src/utils/constants.js
@@ -1,0 +1,2 @@
+export const CDN_LOGO="https://dcassetcdn.com/design_img/1442230/51618/51618_7323624_1442230_c91dc92a_image.jpg";
+export const CDN_IMG="https://res.cloudinary.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_508,h_320,c_fill/";
